+++ src/main/java/com/example/webtoon/exception/GlobalExceptionHandler.java
@@ -1,5 +1,6 @@
 package com.example.webtoon.exception;
 
+import com.example.webtoon.type.ErrorCode;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
