+++ src/main/java/com/example/webtoon/type/ResponseCode.java
@@ -1,4 +1,4 @@
-package com.example.webtoon.payload;
+package com.example.webtoon.type;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
